fix: air interactions, NBT range matching, and JEI tooltip noise
Right-click in the air never fired. onRightClickItem filtered to MAIN_HAND, a
guard copied from onRightClickBlock, but the two events do not behave alike:
Minecraft.startUseItem walks both hands and stops at the first that consumes the
click, so an item held only in the off hand arrives as an OFF_HAND event. Every
"in the air" example is authored with hand "off", so all of them were dropped
before matching. The hand filter is replaced by a per-tick guard, since both
hands can still fire for one click when the main hand passes — and
handlePlayerInteract already inspects both hands itself.

NBT ranges were matched by exact tag class. Authored SNBT mixes widths freely
({RangeTag:[0,500]} is ints, {RangeTag:[2s,7s]} is shorts) and the value read off
the item need not agree, so a valid whitelist could silently never match, or
throw ClassCastException when the two widths differed. Matching is now numeric on
both sides, in the whitelist and blacklist paths alike.

ItemView also wrote enchantment levels as ints where 1.20.1 stored shorts, which
broke every existing `lvl:{RangeTag:[2s,7s]}` whitelist — exactly the format
parity the port was supposed to preserve. It writes shorts again.

The same class-strict read made JEI drop tooltip lines: RangeTag was read as
TAG_INT for plain values and TAG_SHORT for enchantment levels, so whichever width
the file did not use rendered nothing at all. Both now go through one
width-agnostic reader.

Finally, getRegistryName returned the interaction id, which makes JEI attach its
OutputSlotTooltipCallback and append "Recipe By: PunchThemAll" to every drop.
Returning null drops that line; the id is already on the arrow tooltip.

+++ src/main/java/com/drimoz/punchthemall/core/util/ItemView.java
@@ -48,7 +48,9 @@ public final class ItemView {
                 Holder<Enchantment> holder = entry.getKey();
                 CompoundTag e = new CompoundTag();
                 holder.unwrapKey().ifPresent(key -> e.putString("id", key.location().toString()));
-                e.putInt("lvl", entry.getIntValue());
+                // 1.20.1 stored enchantment levels as shorts, and authored files write them as `3s`.
+                // Keep the same width so existing whitelists/blacklists read identically.
+                e.putShort("lvl", (short) entry.getIntValue());
                 list.add(e);
             }
             view.put("Enchantments", list);
